initial implementation

+++ contracts/libraries/FixedPoint.sol
@@ -5,7 +5,6 @@ pragma abicoder v2;
 import "./FullMath.sol";
 
 library FixedPoint {
-
     struct uq112x112 {
         uint224 _x;
     }
@@ -23,22 +22,21 @@ library FixedPoint {
         return uint112(self._x >> RESOLUTION);
     }
 
-    function decode112with18(uq112x112 memory self) internal pure returns (uint) {
-
-        return uint(self._x) / 5192296858534827;
+    function decode112with18(uq112x112 memory self) internal pure returns (uint256) {
+        return uint256(self._x) / 5192296858534827;
     }
 
     function fraction(uint256 numerator, uint256 denominator) internal pure returns (uq112x112 memory) {
-        require(denominator > 0, 'FixedPoint::fraction: division by zero');
+        require(denominator > 0, "FixedPoint::fraction: division by zero");
         if (numerator == 0) return FixedPoint.uq112x112(0);
 
         if (numerator <= uint144(-1)) {
             uint256 result = (numerator << RESOLUTION) / denominator;
-            require(result <= uint224(-1), 'FixedPoint::fraction: overflow');
+            require(result <= uint224(-1), "FixedPoint::fraction: overflow");
             return uq112x112(uint224(result));
         } else {
             uint256 result = FullMath.mulDiv(numerator, Q112, denominator);
-            require(result <= uint224(-1), 'FixedPoint::fraction: overflow');
+            require(result <= uint224(-1), "FixedPoint::fraction: overflow");
             return uq112x112(uint224(result));
         }
     }
